SteeringController veglegesitese
Ami eddig TestSteering-kent futott, atkerult a "rendes" SteeringController.cs-be. A regi SteeringController maradekat meghagytam TestSteeringController-kent, mert meg dolgozunk egy jobb implementacion, amihez hasznaljuk
